significant changes
Added Jakub's church based level design,
added a lot of lights, kinda polished the level more.
added a target feature - crosshair changes color to red when pointed directly at a enemy.
Little changes to the Enemy and Weapon script, there was a bug where the first hit to the enemy. wasn't included to the overall damage it fucked the death with particle system, Sometimes enemy was still alive even tho it had -20hp.

Added Kacper's Fitality blood particle.

Created a example Enemy trigger Event. When we trigger this thing Enemy wave starts to spawn in a certain area, we can tweak the settings in the inspector.